Move react elements to react folder, add empty-state component (work in progress).
