+++ landing/templates/react-components.jade
@@ -8,7 +8,14 @@
       p Placeholder text
       
       +playground({name: "select", darkMode: "false"})
-        include ../../lib/select/demo
+        include ../../lib/react/select/demo
+        
+    section 
+      h2#emptystate Empty state
+      p Use this component for onboarding users when they still don’t have any data.
+      
+      +playground({name: "empty-state", darkMode: "false"})
+        include ../../lib/react/empty-state/demo
