+++ lib/react/empty-state/demo.jade
@@ -1,0 +1,5 @@
+div(
+  styleguide="emptyState",
+  title="Users",
+  description="Here you will find all the users."
+)
