+++ lib/react/empty-state/index.styl
@@ -1,0 +1,35 @@
+.emptyState
+  text-align center
+
+  > .emptyStateTitle
+    font-weight 500
+    margin-bottom 40px
+    font-size 32px
+
+  > .emptyStateImage
+    width 140px
+    height 140px
+    background-color rgba(black, 0.03)
+    border-radius 50%
+    position relative
+    margin 0 auto
+    margin-bottom 20px
+
+    > svg
+      width 70px
+      height 70px
+      color black
+      opacity 0.4
+      position absolute
+      top 50%
+      left 50%
+      transform translate(-50%, -50%)
+
+  > .emptyStateDescription
+    display block
+    max-width 440px
+    margin 0 auto
+    margin-bottom 60px
+
+  > .emptyStateButtons .btn + .btn
+    margin-left 20px
